+++ packages/terra-form-select/src/Select.jsx
@@ -10,6 +10,11 @@ import Util from './_SelectUtil';
 
 const propTypes = {
   /**
+   * Whether a clear option is available to clear the selection.
+   * This is not applicable to the `multiple` or `tag` variants since the selection can already be deselected using the tag.
+   */
+  allowClear: PropTypes.bool,
+  /**
    * The dropdown menu options.
    */
   children: PropTypes.node,
@@ -91,6 +96,7 @@ const propTypes = {
 };
 
 const defaultProps = {
+  allowClear: false,
   children: undefined,
   defaultValue: undefined,
   disabled: false,
@@ -197,7 +203,7 @@ class Select extends React.Component {
   render() {
     const { intl } = this.context;
     const {
-      children, defaultValue, onChange, placeholder, value, ...otherProps
+      allowClear, children, defaultValue, onChange, placeholder, value, ...otherProps
     } = this.props;
 
     const defaultPlaceholder = intl.formatMessage({ id: 'Terra.form.select.defaultDisplay' });
@@ -211,6 +217,7 @@ class Select extends React.Component {
         onDeselect={this.handleDeselect}
         onSelect={this.handleSelect}
         placeholder={selectPlaceholder}
+        clearOptionDisplay= {allowClear ? selectPlaceholder.length > 0 ? selectPlaceholder : defaultPlaceholder : undefined}
         dropdown={dropdownProps => (
           <DropdownMenu intl={intl} {...dropdownProps}>
             {this.state.tags}