+++ packages/terra-form-select/tests/jest/Menu.test.jsx
@@ -56,4 +56,24 @@ describe('Menu', () => {
     expect(wrapper).toMatchSnapshot();
     expect(liveRegion.current.innerText).toEqual('No matching results for "asdf"');
   });
+
+  it('should render a Menu with the clear option', () => {
+    const liveRegion = { current: document.createElement('div') };
+
+    jest.useFakeTimers();
+
+    const menu = (
+      <Menu onSelect={() => {}} visuallyHiddenComponent={liveRegion} variant="combobox" clearOptionDisplay="-Select-" searchValue="">
+        <Option value="value" display="display" />
+      </Menu>
+    );
+
+    jest.useFakeTimers();
+
+    const wrapper = shallow(menu, intlContexts.shallowContext);
+
+    jest.runOnlyPendingTimers();
+
+    expect(wrapper).toMatchSnapshot();
+  });
 });