+++ packages/terra-form-select/CHANGELOG.md
@@ -10,6 +10,11 @@ Unreleased
 ### Fixed
 * Fixed issue preventing the input from inheriting the base font family
 
+5.6.0 - (February 12, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 5.5.0 - (February 5, 2019)
 ------------------
 ### Added